use env

+++ prisma/migrations/20230716190540_init/migration.sql
@@ -1,6 +1,6 @@
 -- CreateTable
 CREATE TABLE `User` (
-    `_id` INTEGER NOT NULL,
+    `_id` DOUBLE NOT NULL,
     `email` VARCHAR(191) NOT NULL,
     `name` VARCHAR(191) NULL,
 
@@ -10,14 +10,14 @@ CREATE TABLE `User` (
 
 -- CreateTable
 CREATE TABLE `Post` (
-    `_id` INTEGER NOT NULL,
+    `_id` DOUBLE NOT NULL,
     `createdAt` DATETIME(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3),
     `updatedAt` DATETIME(3) NOT NULL,
     `title` VARCHAR(191) NOT NULL,
     `content` VARCHAR(191) NULL,
     `published` BOOLEAN NOT NULL DEFAULT false,
     `viewCount` INTEGER NOT NULL DEFAULT 0,
-    `authorId` INTEGER NULL,
+    `authorId` DOUBLE NULL,
 
     PRIMARY KEY (`_id`)
 ) DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci;
